Install refactoring.nvim in development mode.
Adds refactoring.nvim from the local disk to the project. Configure nvim
to use the shortcuts for the existing refactorings to re, rt and rf.

The refactoring commands are not working yet, opened a ticket in
https://github.com/ThePrimeagen/refactoring.nvim/issues/32 to find out
whats wrong with the installation.
Errors in the refactoring.nvim are only when actually using the command
for refactoring, so we leave this broken refactoring.nvim installation
in - it does not bother me now.